pyproject.toml
[project]
name = "translater"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ell-ai[all]>=0.0.15",
    "icecream>=2.1.3",
    "langchain-text-splitters>=0.3.2",
    "pydantic-settings>=2.6.1",
    "regex>=2024.11.6",
    "tiktoken>=0.8.0",
]
